requirements.txt
requests==2.31.0
geopandas==0.14.1
shapely==2.0.2
pandas==2.1.3
matplotlib==3.8.2
python-dotenv==1.0.0
pytest==7.4.3
pytest-mock==3.12.0
fiona==1.9.6